+++ src/pages/OrderPage/OrderPage.tsx
@@ -6,19 +6,11 @@ import SenderSectionComponent from "./components/SenderSection/SenderSection";
 import ReceiverSectionComponent from "./components/ReceiverSection/ReceiverSection";
 import ProductInfo from "./components/ProductInfo/ProductInfo";
 import { useProductInfo } from "./hooks/useProductInfo";
-import { ROUTES } from "@/constants/routes";
-import { useNavigate } from "react-router-dom";
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import { useOrderForm } from "./hooks/useOrderForm";
-import { ORDER_MESSAGES } from "./constants/alert";
 import { LoadingSpinner } from "@/components/common/LoadingSpinner";
 import { LoadingContainer } from "../HomePage/components/Category/Category.styles";
-import { useAuth } from "@/contexts/AuthContext";
-import type { Order } from "@/types/Order";
-import { createOrder } from "@/data/api";
-import { AxiosError } from "axios";
-import { useFetch } from "@/hooks/useFetch";
-import { API_ERROR_MESSAGES } from "./constants/apiMessage";
+import { useOrderSubmit } from "./hooks/useOrderSubmit";
 
 const OrderPageContainer = styled.div`
   display: flex;
@@ -28,36 +20,8 @@ const OrderPageContainer = styled.div`
 `;
 
 function OrderPage() {
-  const navigate = useNavigate();
   const [isSubmittedOnce, setIsSubmittedOnce] = useState(false);
   const { product, loading } = useProductInfo();
-  const { user } = useAuth();
-
-  const [order, setOrder] = useState<Order | null>(null);
-
-  const { data } = useFetch<{ success: boolean }>({
-    fetchFn: () => {
-      if (!order) {
-        return Promise.reject(new Error(API_ERROR_MESSAGES.ORDER_NOT_FOUND));
-      }
-      return createOrder(user?.authToken || "", order);
-    },
-    errorHandler: (error) => {
-      if (error instanceof AxiosError) {
-        if (error.response?.status === 401) {
-          navigate(ROUTES.LOGIN);
-        }
-      }
-    },
-    enabled: !!order, // order가 있을 때만 실행
-    deps: [order], // order가 변경되면 자동 실행
-  });
-
-  useEffect(() => {
-    if (data && data.success) {
-      navigate(ROUTES.HOME);
-    }
-  }, [data, navigate]);
 
   const {
     messageCard,
@@ -75,43 +39,14 @@ function OrderPage() {
     getFormValues,
   } = useOrderForm({ isSubmittedOnce });
 
-  const onSubmitHandler = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const isValid = await validateAllForms();
-
-    if (!isValid) {
-      setIsSubmittedOnce(true);
-    }
-
-    if (isValid) {
-      const formValues = getFormValues();
-      alert(
-        ORDER_MESSAGES.ORDER_COMPLETE_TEMPLATE({
-          productName: product?.name || "",
-          totalQuantity: formValues.totalQuantity,
-          senderName: formValues.senderName,
-          cardMessage: formValues.cardMessage,
-        })
-      );
-
-      const transformedReceivers = receivers.map((receiver) => ({
-        ...receiver,
-        quantity: Number(receiver.quantity),
-      }));
-
-      const orderData: Order = {
-        productId: product?.id || 0,
-        message: formValues.cardMessage,
-        messageCardId: messageCard.id.toString(),
-        ordererName: formValues.senderName,
-        receivers: transformedReceivers,
-      };
-
-      setOrder(orderData);
-      return;
-    }
-  };
+  const { onSubmitHandler } = useOrderSubmit({
+    validateAllForms,
+    setIsSubmittedOnce,
+    getFormValues,
+    receivers,
+    messageCard,
+    product,
+  });
 
   if (loading) {
     return (
